Consumer/requirements.txt
kafka-python
pymongo==4.4.0